Make FileSystem's field `Backup` a No-Op
Backup field is used during creation of an EFS File System,
and nowhere else. For Find and Update, the API instead uses a
BackupPolicy field. Whenever a user tries setting the Backup
field, the controller will return a Terminal Error.

I am choosing sdkFind to return this failure since it's the funtion
that is called after each event that happens to a resource.