Make FileSystem's field `Backup` a No-Op (#29)
Issue [#2182](https://github.com/aws-controllers-k8s/community/issues/2182)

Description of changes:
Backup field is used during creation of an EFS File System, 
and nowhere else. For Find and Update, the API instead uses 
a BackupPolicy field.
Moving forward the ACK controller will return a Terminal error
when the Backup field is defined, and instead we will only be accepting 
backup configurations through the backupPolicy field.

I am choosing sdkFind to return this failure since it's the funtion 
that is called after each event that happens to a resource.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.